scheduler/requirements.txt
clic
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
varname
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
